Move interactive elements to app.py
